fix(ui): merge flow correctness, Email filter, empty-state dead end
Third UX pass — the merge flow, screenshare mode and the empty state, none of
which had been exercised against the running app before.

Merge was broken in two ways:
- "Merge & paste" set the clipboard and sent ⌘V *without hiding the launcher*,
  so the merged text landed in Magpie's own search field instead of the app the
  user came from. It now goes through the same deferred hide → paste path as
  Enter and the slot hotkeys.
- The queue stored list *indices* (merge_set: Vec<i32>), so changing the search
  text or a filter left the ticks pointing at whatever now occupied those
  positions — and Merge & paste silently merged the wrong entries. It stores
  entry ids now; merged_text() already walks the ids in order, so queue order
  is preserved.
- The separator options shaped as tofu (↵ → an empty box, ␣ → "__"): the three
  choices were literally unreadable. Now "New line / Space / Comma" as a
  segmented control, with accent reserved for the one primary action in the row
  — two adjacent blue pills made the selection indistinguishable from the
  button. "Merge 3 · sep:" → "3 items queued · join with".

Email was an unreachable filter: TypeFilter::Email → Kind::Email existed and
the capture layer classifies entries as Email, but no chip and no index mapped
to it. Added the chip, shifted the index mapping, and added a test asserting
every TypeFilter variant is reachable so this can't regress.

Empty state was a dead end:
- The detail pane stayed as an empty dark slab; it's now dropped entirely when
  there are no entries, giving the message the full width.
- The status bar promised "⏎ Paste to X · ⌘C Copy · ⌘K Actions" with nothing
  selected. Those are hidden now.
- Added a "Clear search & filters" button, and clear_filters() now really
  clears everything (type chip and tag included, not just app/time/pinned).
- `filtered` ignored the type chip and tag, so filtering to Image with no
  images said "No clipboard history yet" instead of "No matching items".

Screenshare mode gave no explanation — a wall of dots reads as corrupt data.
Added an inline banner naming the mode and how to leave it, and dropped the 👁
from "Reveal" (it shaped as a blob at 12px).

Dev tooling: MAGPIE_UI_TOUR gains merge / mask / empty / per-type-chip steps,
and TYPE_CHIPS documents the chip order the index mapping depends on.

+++ crates/magpie-app/src/app_state.rs
@@ -7,8 +7,10 @@ pub struct AppState {
     pub store: Mutex<Store>,
     pub images: FsImageStore,
     pub ui: Mutex<UiState>,
-    /// Entry indices (into the current results) queued for merge-paste.
-    pub merge_set: Mutex<Vec<i32>>,
+    /// Entry **ids** queued for merge, in the order they were added. Ids, not
+    /// list indices: the queue has to survive a search or filter change, which
+    /// reorders (and re-populates) the result list under it.
+    pub merge_set: Mutex<Vec<i64>>,
     /// Session-only screensharing mode (masks everything). Not persisted.
     pub screenshare: Mutex<bool>,
     /// Masking config snapshot (from `Config`, immutable for the session).
